block empty lines in text area

+++ modules/client/src/main/scala/vinyldns/client/pages/zoneview/components/RecordSetTable.scala
@@ -198,26 +198,28 @@ object RecordSetTable {
         <.td(
           <.div(
             ^.className := "btn-group",
-            <.button(
-              ^.className := "btn btn-info btn-rounded test-edit",
-              ^.`type` := "button",
-              ^.onClick --> makeUpdateRecordModalVisible(recordSet),
-              ^.title := s"Update record set (name: ${recordSet.name}, id: ${recordSet.id})",
-              ^.disabled := !(recordSet.accessLevel == "Update" || recordSet.accessLevel == "Delete"),
-              VdomAttr("data-toggle") := "tooltip",
-              <.span(^.className := "fa fa-edit"),
-              " Update"
-            ),
-            <.button(
-              ^.className := "btn btn-danger btn-rounded test-delete",
-              ^.`type` := "button",
-              ^.onClick --> deleteRecordSet(P, S, recordSet),
-              ^.title := s"Delete record set (name: ${recordSet.name}, id: ${recordSet.id})",
-              ^.disabled := recordSet.accessLevel != "Delete",
-              VdomAttr("data-toggle") := "tooltip",
-              <.span(^.className := "fa fa-trash"),
-              " Delete"
-            )
+            if (recordSet.canUpdate(P.zone.name))
+              <.button(
+                ^.className := "btn btn-info btn-rounded test-edit",
+                ^.`type` := "button",
+                ^.onClick --> makeUpdateRecordModalVisible(recordSet),
+                ^.title := s"Update record set (name: ${recordSet.name}, id: ${recordSet.id})",
+                VdomAttr("data-toggle") := "tooltip",
+                <.span(^.className := "fa fa-edit"),
+                " Update"
+              )
+            else TagMod.empty,
+            if (recordSet.canDelete(P.zone.name))
+              <.button(
+                ^.className := "btn btn-danger btn-rounded test-delete",
+                ^.`type` := "button",
+                ^.onClick --> deleteRecordSet(P, S, recordSet),
+                ^.title := s"Delete record set (name: ${recordSet.name}, id: ${recordSet.id})",
+                VdomAttr("data-toggle") := "tooltip",
+                <.span(^.className := "fa fa-trash"),
+                " Delete"
+              )
+            else TagMod.empty
           )
         )
       )
